cmake: modules: dts: Expose DTS_INCLUDE_FILES
This was previously used as a variable internal to the module, but it
can be used in other places in the build system, which may want a list
of DT-specific includes. Therefore, document it as an output variable.

(cherry picked from commit 22fd74914a033f28c954312253a9cf49001d305f)

Original-Signed-off-by: Grzegorz Swiderski <grzegorz.swiderski@nordicsemi.no>
GitOrigin-RevId: 22fd74914a033f28c954312253a9cf49001d305f
Change-Id: I5f9d20a48d22b63f5faf6d8a37767451fd150aaa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5534894
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>