Bluetooth: samples: Switch to one-time adv
This patch removes all uses of the adv auto-resume feature in the
Bluetooth samples. The auto-resume feature is planned for deprecation.

Samples that are not intended to demonstrate a technique to do with
re-connection simply do not restart the advertiser, in interest of
simplicity. The user is expected to reboot the sample when needed.

(cherry picked from commit c6ad4a792724ed19e51c0a0d0f010490150d4eff)

Original-Signed-off-by: Aleksander Wasaznik <aleksander.wasaznik@nordicsemi.no>
GitOrigin-RevId: c6ad4a792724ed19e51c0a0d0f010490150d4eff
Change-Id: I7ca0179bedadea0b46dacb23c631153dd5684106
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5583691
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>