tests: random: Move and rename rng test
Random number generator is a subsystem and although it is mostly used
in crypto, this is not its only utility.

Move the current random number generator test to it is own space
(tests/subsys/random) and rename it to rng.

We need more and better tests for rng, this is an initial commit
to organize it and get ready for further tests.

(cherry picked from commit 8c5806472e4723e6c2e27a65bdaca454c62b7c76)

Original-Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>
GitOrigin-RevId: 8c5806472e4723e6c2e27a65bdaca454c62b7c76
Change-Id: I38b522d69d8fb9dfcc8c895599b53816665ba7ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5588587
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>