twister: fix serial connection for flash_before
The --flash-before flag allows devices to be flashed before the serial
connection is established. However, the implementation was incomplete and
only worked if the port address already existed at the start of the run.
This is incompatible with devices that implement the USB in software
(eg: USB-CDC).

This commit fixes the implementation to delay setting up the connection
until after the device is flashed, and to retry the connection for two
seconds after flashing to give the device time to enumerate as a USB
device.

(cherry picked from commit 7eaca455faf6ff8d9a92c4392da281dab08d0c2b)

Original-Signed-off-by: Mike Szczys <mike@golioth.io>
GitOrigin-RevId: 7eaca455faf6ff8d9a92c4392da281dab08d0c2b
Change-Id: Ieb774f9521d3cfd8e5e02547c73d950aba864bfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5563018
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>