drivers: mipi_dbi: smartbond: Optimize driver PM
This commit should optimize the way the device is allowed
to enter the suspended state. Instead of returning a PM
error code to abort the PM process, the standby power state
is constrained as long as the device is not allowed to enter
suspension. With that approach, acquiring PD_SYS is not needed
when in PM device runtime mode.

(cherry picked from commit 2958f691bb508e642a41b731f7df9d31ee8500e0)

Original-Signed-off-by: Ioannis Karachalios <ioannis.karachalios.px@renesas.com>
GitOrigin-RevId: 2958f691bb508e642a41b731f7df9d31ee8500e0
Change-Id: I07072b33a1d1165963ae528e4b300f27dc05b174
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5570550
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>