drivers: display: smartbond: Optimize driver PM
This commit should optimize the way the device is allowed
to enter the suspended state. Instead of returning a PM
error code to abort the PM process, the standby power state
is constrained as long as the device is not allowed to enter
suspension. With that approach, acquiring PD_SYS is not needed
when in PM device runtime mode.

(cherry picked from commit 71a5f1b9fdcd14c2b766d758031d25e127e939a0)

Original-Signed-off-by: Ioannis Karachalios <ioannis.karachalios.px@renesas.com>
GitOrigin-RevId: 71a5f1b9fdcd14c2b766d758031d25e127e939a0
Change-Id: Ib0d36457bd80f4e2138d3b3232bbcc0788bb727e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5570551
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>