drivers: entropy: smartbond: Optimize driver PM
This commit should optimize the way the device is allowed
to enter the suspended state. Instead of returning a PM
error code to abort the PM process, the standby power state
is constrained as long as the device is not allowed to enter
suspension. With that approach, acquiring PD_SYS is not needed
when in PM device runtime mode.

(cherry picked from commit 62f8f160ada4a3239217ff9ea00dfb489623fdaf)

Original-Signed-off-by: Ioannis Karachalios <ioannis.karachalios.px@renesas.com>
GitOrigin-RevId: 62f8f160ada4a3239217ff9ea00dfb489623fdaf
Change-Id: I84671d60417eedbde92f2530a6d4eaa8edafab33
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5570552
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>