soc: boards: Add Mediatek MT8186 and MT8188 audio DSPs
These are very similar devices to mt8195, minimal changes needed
beyond boilerplate configuration.

In the process, this reworks the board/soc layout to be HWMv2
compliant, with "adsp" becoming a CPU cluster beneath the SOC.  So the
name of the boards to west become e.g. "mt8195/mt8195/adsp" (which can
be shortened to "mt8195//adsp" if desired).

Note that the cpuclk driver is not yet ported, it works only with 8195
(the clocking/power architecture seems similar between the parts, but
the graph of wells and clocks is different and historically these have
been three separate drivers in SOF).  The biggest changes are in the
image/loader scripts, which needed some rework for cross-device
portability.

(cherry picked from commit 5364783ba181825d2095785fb56abb3944d9dc90)

Original-Signed-off-by: Andy Ross <andyross@google.com>
GitOrigin-RevId: 5364783ba181825d2095785fb56abb3944d9dc90
Cr-Build-Id: 8729941316818268033
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8729941316818268033
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ibbc02facc57a10287bef9c35fb210fd68f469cec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6059559
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>