net: ipv6: scope checking function fix
Change the implementation of net_ipv6_is_addr_mcast_scope() inline
function that let us check if a given IPv6 address has a specified
scope. Previously, it was comparing the whole byte including flags
of a multicast address. It meant, that while checking for a specific
scope a one was also checking the flags. Even in Zephyr's net stack
there are checks for a IPv6 link local scope that are failing for
addresses that are not marked as "well known" (when least significant
bit of the flags is set).

(cherry picked from commit 3da35039ceb30d80e689d058626c2222142f16f0)

Original-Signed-off-by: Konrad Derda <konrad.derda@nordicsemi.no>
GitOrigin-RevId: 3da35039ceb30d80e689d058626c2222142f16f0
Change-Id: If0f661869c8e47db69093a2867a155c69f83b961
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5575346
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>