storage: flash_map: replace TinyCrypt by PSA
As part of ongoing work to move away from TinyCrypt and towards PSA
(#43712), introduce a PSA option and remove the TinyCrypt one for the
SHA-256 implementation.

The Mbed TLS implementation is modified to use `mbedtls_sha256`
directly for smaller code size.

As of now the implementation defaults to PSA only if TF-M is enabled
because a dependency loop happens if using `PSA_CRYPTO_CLIENT` as a
condition in `FLASH_AREA_CHECK_INTEGRITY_BACKEND`.

A test case is added for the PSA implementation, and an NS platform is
added to the base test case to verify the compilation on a TF-M-enabled
platform.

(cherry picked from commit f2c643ba0205c00dccea56af5294c60d1e92aea3)

Original-Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>
GitOrigin-RevId: f2c643ba0205c00dccea56af5294c60d1e92aea3
Change-Id: I6cad6c979237fab3dbb287bed9a89b5cd5f05972
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5635327
Commit-Queue: Dawid Niedźwiecki <dawidn@google.com>
Tested-by: Dawid Niedźwiecki <dawidn@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>