Bluetooth: Host: Rework enabling of scanner
To make the scanner module more understandable and more streamlined, I
reworked the update mechanism of the scanner. The scanner tracks now the
parameters that were used to enable it and the reason why it is running.
This facilitates state logic and allows other modules to "start the
scanner", altough it is already running.
This is mostly a refactoring and not a functional change.

Added a test to verify the behavior.

add to task

(cherry picked from commit c910520f4d047ab395b705d56d1a3a2809b8cf69)

Original-Signed-off-by: Jan Müller <jan.mueller@nordicsemi.no>
GitOrigin-RevId: c910520f4d047ab395b705d56d1a3a2809b8cf69
Cr-Build-Id: 8739001010820476625
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8739001010820476625
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ia2ee475fd967c7dcf2957734065cc0d71311dc3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5803450
Reviewed-by: Eric Yilun Lin <yllin@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Eric Yilun Lin <yllin@google.com>
Commit-Queue: Eric Yilun Lin <yllin@google.com>