Bluetooth: Host: Allow conn create timeout longer than RPA timeout
https://github.com/zephyrproject-rtos/zephyr/pull/72674 fixed
a bug where this configuration did not work.

Now that this configuration is tested, we should mark it
as supported.

The timeout check that was present in the code before
was useless and was not working because the check was
run before a default timeout of 0 was converted to a timeout.

(cherry picked from commit 9cf6839b182fdcf0b44d4a117fd57c0d447ff7ce)

Original-Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>
GitOrigin-RevId: 9cf6839b182fdcf0b44d4a117fd57c0d447ff7ce
Change-Id: Ice020d5825447400df12ccbd9be29fc9af00a7a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5625521
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>