doc: develop: tools: add VS Code setup guide
This basic guide only covers linting so far. It could be expanded in the
future to be more accurate and detailed. The current configuration
allows to navigate the code but some symbols are not recognized by the
linter, and the native compiler is always used.

(cherry picked from commit dc987267ac3299284772662d7cf5da7b69e4bad4)

Original-Co-authored-by: Enguerrand de Ribaucourt <enguerrand.de-ribaucourt@savoirfairelinux.com>
Original-Co-authored-by: Luc BEAUFILS <luc.beaufils@savoirfairelinux.com>
Original-Signed-off-by: Enguerrand de Ribaucourt <enguerrand.de-ribaucourt@savoirfairelinux.com>
Original-Signed-off-by: Luc BEAUFILS <luc.beaufils@savoirfairelinux.com>
GitOrigin-RevId: dc987267ac3299284772662d7cf5da7b69e4bad4
Change-Id: I2967ebc6b84706f08964b07995f84645b660edef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5658655
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>