usb: device_next: Clarify UAC2 send buffer requirements
The original intention was that the audio buffers supplied by
application to the UAC2 API can be directly used by UDC driver.
The buffer alignment requirement was explicitly stated with regards
to the receive buffer, but was mistakenly omitted on transmit buffer.

While the missing comment is added, mention also that the buffer has to
be otherwise suitable for use by UDC driver, because alignment might not
be the only UDC driver requirement. For example, the UDC driver can
require the buffer to reside in memory region that it can access with
DMA.

(cherry picked from commit 42c9f79e4e4f042a93d899344ab50373790ccc1f)

Original-Signed-off-by: Tomasz Moń <tomasz.mon@nordicsemi.no>
GitOrigin-RevId: 42c9f79e4e4f042a93d899344ab50373790ccc1f
Change-Id: I8b2a0ccbba8e2de0dedf477f04649931ba71c098
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5629144
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>